moved patternscanner.go to the parent testutil pkg (used not only by containers)
