+++ pkg/network/usm/sharedlibraries/testutil/testutil.go
@@ -21,7 +21,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/network/protocols/http/testutil"
 	usmtestutil "github.com/DataDog/datadog-agent/pkg/network/usm/testutil"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
-	dockerutils "github.com/DataDog/datadog-agent/pkg/util/testutil/docker"
+	procutil "github.com/DataDog/datadog-agent/pkg/util/testutil"
 )
 
 // mutex protecting build process
@@ -31,7 +31,7 @@ var mux sync.Mutex
 // handle to the given paths.
 func OpenFromProcess(t *testing.T, programExecutable string, paths ...string) (*exec.Cmd, error) {
 	cmd := exec.Command(programExecutable, paths...)
-	patternScanner := dockerutils.NewScanner(regexp.MustCompile("awaiting signal"), make(chan struct{}, 1))
+	patternScanner := procutil.NewScanner(regexp.MustCompile("awaiting signal"), make(chan struct{}, 1))
 	cmd.Stdout = patternScanner
 	cmd.Stderr = patternScanner
